pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corrections-agent"
version = "0.1.0"
description = "Corrections Agent services + frontend"
readme = "docs/architecture.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "google-genai>=0.3.0",
  "faiss-cpu>=1.7.4",
  "langchain>=0.1.0",
  "langchain-core>=0.1.0",
  "langchain-community>=0.0.10",
  "langchain-google-genai>=0.0.6",
  "tqdm>=4.66.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "httpx>=0.25.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["services*", "mock_apis*", "eval*", "train*"]
